% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLML.R
\name{MLML}
\alias{MLML}
\title{MLE (maximum likelihood estimates) of 5-mC and 5-hmC levels.}
\usage{
MLML(G = NULL, H = NULL, L = NULL, M = NULL, T = NULL, U = NULL,
  tol = 1e-04, exact = FALSE)
}
\arguments{
\item{G}{Unmethylated channel (intensities/counts) from TAB-conversion (5-C + 5-mC).}

\item{H}{Methylated channel (intensities/counts) from TAB-conversion  (True 5-hmC).}

\item{L}{Unmethylated channel (intensities/counts) from oxBS-conversion (5-C + 5-hmC).}

\item{M}{Methylated channel (intensities/counts) from oxBS-conversion (True 5-mC).}

\item{T}{Methylated channel (intensities/counts) from standard BS-conversion (5-mC+5-hmC).}

\item{U}{Unmethylated channel (intensities/counts) from standard BS-conversion (True 5-C).}

\item{tol}{convergence tolerance; ignored if exact=TRUE}

\item{exact}{logical indicating if the exact constrained MLE should be returned as estimate;
ignored when data is available for all the 3 methods: TAB, oxBS and BS.
If exact=FALSE EM-algorithm is used.}
}
\value{
The returned value is a list with the following components.

\item{mC}{maximum likelihood estimate for the proportion of methylation.}

\item{hmC}{maximum likelihood estimate for the proportion of hydroxymethylation.}

\item{C}{maximum likelihood estimate for the proportion of unmethylation.}

\item{methods}{the conversion methods used to produce the MLE}
}
\description{
MLE (maximum likelihood estimates) of 5-mC and 5-hmC levels.
}
\details{
The function returns MLE estimates based on EM-algorithm for any
 combination of two methods or all three methods. When only two methods are combined, the user
 can set the option exact=TRUE to obtain exact constrained MLE.
 The function assumes that the order of the rows and columns in the input matrices are consistent.
 In addition, all the input matrices must have the same dimension.
 Usually, rows represent CpG loci and columns are the samples.
}
\examples{
library(MLML2R)
# load the example datasets from BS, oxBS and TAB methods
data(MethylatedBS_sim)
data(MethylatedOxBS_sim)
data(UnMethylatedBS_sim)
data(UnMethylatedOxBS_sim)
data(MethylatedTAB_sim)
data(UnMethylatedTAB_sim)

# obtain MLE via EM-algorithm for BS+oxBS:
results_em <- MLML(T = MethylatedBS_sim , U = UnMethylatedBS_sim,
L = UnMethylatedOxBS_sim, M = MethylatedOxBS_sim,tol=0.0001)

# obtain constrained exact MLE for BS+oxBS:
results_exact <- MLML(T = MethylatedBS_sim , U = UnMethylatedBS_sim,
L = UnMethylatedOxBS_sim, M = MethylatedOxBS_sim,exact=TRUE)

# obtain MLE via EM-algorithm for BS+TAB:
results_em <- MLML(T = MethylatedBS_sim , U = UnMethylatedBS_sim,
G = UnMethylatedTAB_sim, H = MethylatedTAB_sim,tol=0.0001)

# obtain constrained exact MLE for BS+TAB:
results_exact <- MLML(T = MethylatedBS_sim , U = UnMethylatedBS_sim,
G = UnMethylatedTAB_sim, H = MethylatedTAB_sim,exact=TRUE)

# obtain MLE via EM-algorithm for oxBS+TAB:
results_em <- MLML(L = UnMethylatedOxBS_sim, M = MethylatedOxBS_sim,
G = UnMethylatedTAB_sim, H = MethylatedTAB_sim,tol=0.0001)

# obtain constrained exact MLE for oxBS+TAB:
results_exact <- MLML(L = UnMethylatedOxBS_sim, M = MethylatedOxBS_sim,
G = UnMethylatedTAB_sim, H = MethylatedTAB_sim,exact=TRUE)

# obtain MLE via EM-algorithm for BS+oxBS+TAB:
results_em <- MLML(T = MethylatedBS_sim , U = UnMethylatedBS_sim,
L = UnMethylatedOxBS_sim, M = MethylatedOxBS_sim,
G = UnMethylatedTAB_sim, H = MethylatedTAB_sim,tol=0.0001)

}
\references{
Qu J, Zhou M, Song Q, Hong EE, Smith AD. MLML: consistent simultaneous estimates of DNA methylation and hydroxymethylation.
  Bioinformatics. 2013;29(20):2645-2646. doi:10.1093/bioinformatics/btt459.

  Zongli Xu, Jack A. Taylor, Yuet-Kin Leung, Shuk-Mei Ho, Liang Niu;
  oxBS-MLE: an efficient method to estimate 5-methylcytosine and 5-hydroxymethylcytosine
  in paired bisulfite and oxidative bisulfite treated DNA,
  Bioinformatics, 2016;32(23):3667–3669. https://doi.org/10.1093/bioinformatics/btw527
}
\author{
Samara Kiihl samara@ime.unicamp.br;
Maria Jose Garrido;
Arce Domingo-Relloso;
Jose Bermudez;
Maria Tellez-Plaza.
}
