% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLML2R.R
\name{MLML2R}
\alias{MLML2R}
\title{MLE estimates of 5-mC and 5-hmC levels.}
\usage{
MLML2R(G = NULL, H = NULL, L = NULL, M = NULL, T = NULL, U = NULL,
  tol = 0.001)
}
\arguments{
\item{G}{no 5-hmC from TAB-conversion.}

\item{H}{5-hmC from TAB-conversion.}

\item{L}{no 5-mC from oxBS-conversion.}

\item{M}{5-mC from oxBS-conversion.}

\item{T}{5-mC+5-hmC from BS-conversion.}

\item{U}{no methylation from BS-conversion.}

\item{tol}{convergent tolerance.}
}
\value{
Returns a list with estimates for proportion of methylation (mC), proportion of hydroxymethylation (hmC), and proportion of no methylation (C).
}
\description{
MLE estimates of 5-mC and 5-hmC levels.
}
