% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T_OxBS_sim}
\alias{T_OxBS_sim}
\title{T (converted) counts from oxBS}
\format{100 x 2 matrix (CpGs in the rows and  Samples in the columns).}
\usage{
T_OxBS_sim
}
\description{
A matrix of simulated counts corresponding to 100 CpGs and 2 samples.
True proportions of methylation, hydroxymethylation and unmethylated
used in the simulation are .3, .2, and .5, respectively.
}
\keyword{datasets}
