% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_parameters_sim2}
\alias{true_parameters_sim2}
\title{True proportions for simulation}
\format{a data frame with 1000 rows (CpGs) and 3 columns:
\describe{
  \item{p_m}{methylation proportion}
  \item{p_h}{hydroxymethylation proportion}
  \item{p_u}{unmethylated proportion}
}}
\usage{
true_parameters_sim2
}
\description{
A dataset with the true proportions of methylation, hydroxymethylation
and unmethylation for each CpG. Used to generate the counts.
}
\keyword{datasets}
