% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnMethylatedBS}
\alias{UnMethylatedBS}
\title{Unmethylated counts from BS}
\format{100 x 2 matrix (CpGs in the rows and  Samples in the columns).}
\usage{
UnMethylatedBS
}
\description{
A matrix of counts after preprocessing corresponding to 100 CpGs and 2 samples.
}
\keyword{datasets}
