% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLML.R
\name{MLML}
\alias{MLML}
\title{MLE (maximum likelihood estimates) of 5-mC and 5-hmC levels.}
\usage{
MLML(G = NULL, H = NULL, L = NULL, M = NULL, T = NULL, U = NULL,
  tol = 1e-04)
}
\arguments{
\item{G}{Unmethylated channel (intensities/counts) from TAB-conversion (5-C + 5-mC).}

\item{H}{Methylated channel (intensities/counts) from TAB-conversion  (True 5-hmC).}

\item{L}{Unmethylated channel (intensities/counts) from oxBS-conversion (5-C + 5-hmC).}

\item{M}{Methylated channel (intensities/counts) from oxBS-conversion (True 5-mC).}

\item{T}{Methylated channel (intensities/counts) from standard BS-conversion (5-mC+5-hmC).}

\item{U}{Unmethylated channel (intensities/counts) from standard BS-conversion (True 5-C).}

\item{tol}{convergence tolerance.}
}
\value{
The returned value is a list with the following components.

\item{mC}{maximum likelihood estimate for the proportion of methylation.}

\item{hmC}{maximum likelihood estimate for the proportion of hydroxymethylation.}

\item{C}{maximum likelihood estimate for the proportion of unmethylation.}

\item{methods}{the conversion methods used to produce the MLE}
}
\description{
MLE (maximum likelihood estimates) of 5-mC and 5-hmC levels.
}
\references{
Qu J, Zhou M, Song Q, Hong EE, Smith AD. MLML: consistent simultaneous estimates of DNA methylation and hydroxymethylation.
  Bioinformatics. 2013;29(20):2645-2646. doi:10.1093/bioinformatics/btt459.
}
